"Add a file: dict_guess_number.py. Inside 1 function: guess_number(). The application  prompts the user to enter a maximum number and creates a set of possible numbers ranging from 1 to the maximum number entered.
The user is then prompted to enter a list of numbers among which the target number is hidden. If the user needs help, entering "Help!" will print out all the possible numbers.

After the user enters the numbers, they are asked if the target number is among them. If the answer is "Yes", the set of possible numbers is updated to include only those in the given list. If the answer is "No", the set of possible numbers is updated to exclude those in the given list."